+++ @utahdts/utah-design-system/react/components/buttons/Button.jsx
@@ -12,19 +12,19 @@ import { Spinner } from '../widgetsIndicators/Spinner';
 
 /**
  * @param {object} props
- * @param {ButtonAppearance} [props.appearance] the appearance of the button
+ * @param {ButtonAppearance} [props.appearance]
  * @param {import('react').ReactNode} props.children most often is the title of the button, but can also contain most anything
  * @param {string} [props.className] modify your button via className like 'button--primary' and other modifiers found in the button.scss
  * @param {ComponentColors} [props.color] the base color of the button
  * @param {import('react').RefObject<HTMLButtonElement | null>} [props.innerRef] a ref to attach to the actual DOM <button> element
- * @param {import('react').ReactNode} [props.iconLeft] an icon for the left/right side
+ * @param {import('react').ReactNode} [props.iconLeft]
  * @param {import('react').ReactNode} [props.iconRight]
- * @param {boolean} [props.isDisabled] button isDisabled state
- * @param {string} [props.id] the button id
+ * @param {string} [props.id]
  * @param {boolean} [props.isBusy] if the button is busy then it shows a spinner indicator on it and disables the button
- * @param {import('react').MouseEventHandler} props.onClick event for when the button is clicked: (e) => { ... do something with e ...}
+ * @param {boolean} [props.isDisabled]
+ * @param {import('react').MouseEventHandler<HTMLButtonElement>} props.onClick
  * @param {FormElementSizes} [props.size]
- * @param {ButtonTypes} [props.type] button type
+ * @param {ButtonTypes} [props.type]
  * @returns {import('react').JSX.Element}
  */
 export function Button({
@@ -50,7 +50,7 @@ export function Button({
         className,
         `button--${appearance}`,
         // default color is none
-        (color && color !== 'none') ? `button--${color}-color` : null,
+        (color && color !== componentColors.NONE) ? `button--${color}-color` : null,
         // default size is medium
         (size && size !== formElementSizesEnum.MEDIUM) ? `button--${size}` : null
       )}
@@ -58,16 +58,13 @@ export function Button({
       id={id}
       onClick={handleEvent((e) => onClick?.(e))}
       ref={innerRef}
-      // eslint-disable-next-line react/button-has-type
       type={type}
       {...rest}
     >
-      {
-        iconLeft
-          ? <span className="button--icon button--icon-left">{iconLeft}</span>
-          : null
-      }
+      {iconLeft ? <span className="button--icon button--icon-left">{iconLeft}</span> : null}
+
       {children}
+
       {
         isBusy
           ? (
@@ -79,11 +76,8 @@ export function Button({
           )
           : null
       }
-      {
-        iconRight
-          ? <span className="button--icon button--icon-right">{iconRight}</span>
-          : null
-      }
+
+      {iconRight ? <span className="button--icon button--icon-right">{iconRight}</span> : null}
     </button>
   );
 }