+++ @utahdts/utah-design-system/react/components/buttons/SegmentedButton.jsx
@@ -1,3 +1,0 @@
-export function SegmentedButton() {
-  return null;
-}