+++ CHANGELOG.md
@@ -5,13 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [3.0.4] 02/06/2025
+## Fixed
+- Fixed the banner icon + text alignment
+- Fixed text in various places on the website to address clarity and spelling
+- Fixed onChange/onClear types for Select onChange
+
+## Changed
+- HTML ids were updated to be consistent across the components
+
+## Update
+- Added clarity for the Utah Header logo and color customization options
+- Added a better example of a Select component
+- Removed mention of Form Context functionality
+
+
 # [3.0.3] 12/19/2024
 ## Fixed
 - Prevent infinite looping on remove multi select tag
 
 # [3.0.2] 12/10/2024
 ## Fixed
-- Readded './css' export for css path
+- Re-added './css' export for css path
 
 # [3.0.1] 12/2/2024
 ## Fixed